components: refactor tests and components
Extract Image component from Logo and add test to component, add
cardClasses and fix card styles.
